Extract common parts into base render strategy class
Part of #227108
